requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.1
pydantic_core==2.23.3
PySocks==1.7.1
pytest==8.3.2
pytest-dotenv==0.5.2
python-dotenv==1.0.1
requests==2.32.3
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
python-decouple==3.8
psycopg2==2.9.9
psycopg2-binary==2.9.9